Rename media_player to base
